Move example target and clean to unit test
