+++ tests/basic.test.ts
@@ -4,11 +4,13 @@ import {
   runAlways,
   task,
   TrackedFile,
+  trackFile,
 } from "../dnit.ts";
 
 import { assertEquals } from "https://deno.land/std@0.117.0/testing/asserts.ts";
 
 import { Manifest } from "../manifest.ts";
+import { path } from "../deps.ts";
 
 Deno.test("basic test", async () => {
   const tasksDone: { [key: string]: boolean } = {};
@@ -136,4 +138,61 @@ Deno.test("async file deps test", async () => {
 
   assertEquals(tasksDone["taskA"], true);
   assertEquals(tasksDone["taskB"], true);
+});
+
+
+
+Deno.test("tasks with target and clean", async () => {
+
+  const tempDir = await Deno.makeTempDir();
+
+  console.log('tempDir', tempDir)
+
+  const exampleTarget1 = trackFile({
+    path: path.join(tempDir, 'exampleTarget1.txt'),
+  });
+  const testTask1 = task({
+    name: 'testTask1',
+    description: 'Test task to generate (and allow clean) of an example target',
+    action: async () => {
+      await Deno.writeTextFile(exampleTarget1.path, 'example contents');
+    },
+    targets: [exampleTarget1],
+  });
+
+  const exampleTarget2 = trackFile({
+    path: path.join(tempDir, 'exampleTarget2.txt'),
+  });
+  const testTask2 = task({
+    name: 'testTask2',
+    description: 'Test task to generate (and allow clean) of an example target',
+    action: async () => {
+      await Deno.writeTextFile(exampleTarget2.path, 'example contents');
+    },
+    targets: [exampleTarget2],
+  });
+
+  // precheck nonexists
+  assertEquals(await exampleTarget1.exists(), false);
+  assertEquals(await exampleTarget2.exists(), false);
+
+  // setup exec ctx
+  const ctx = await execBasic([], [testTask1, testTask2], new Manifest(""));
+
+  // run test tasks
+  await ctx.getTaskByName("testTask1")?.exec(ctx);
+  assertEquals(await exampleTarget1.exists(), true);
+
+  await ctx.getTaskByName("testTask2")?.exec(ctx);
+  assertEquals(await exampleTarget2.exists(), true);
+
+  // clean
+  await ctx.getTaskByName("clean")?.exec(ctx);
+
+  // check nonexists
+  assertEquals(await exampleTarget1.exists(), false);
+  assertEquals(await exampleTarget2.exists(), false);
+
+  // clean tempdir
+  await Deno.remove(tempDir, { recursive: true });
 });
